Refactor: V2를 V1으로 변경 및 Test 코드 전부 하나로 migration
