+++ src/main/java/com/posomo/saltit/service/RestaurantServiceV1.java
@@ -1,10 +1,11 @@
 package com.posomo.saltit.service;
 
-import org.springframework.data.domain.Slice;
+import org.springframework.stereotype.Service;
 
 import com.posomo.saltit.domain.exception.NoRecordException;
 import com.posomo.saltit.domain.restaurant.dto.RestaurantDetailResponse;
 import com.posomo.saltit.domain.restaurant.dto.RestaurantFilterRequest;
+import com.posomo.saltit.domain.restaurant.dto.RestaurantSearchCondition;
 import com.posomo.saltit.domain.restaurant.dto.RestaurantSummaryResponse;
 import com.posomo.saltit.domain.restaurant.entity.Restaurant;
 import com.posomo.saltit.repository.RestaurantRepository;
@@ -12,7 +13,7 @@ import com.posomo.saltit.repository.RestaurantRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 
-// @Service
+@Service
 @RequiredArgsConstructor
 @Transactional
 public class RestaurantServiceV1 implements RestaurantService {
@@ -20,19 +21,22 @@ public class RestaurantServiceV1 implements RestaurantService {
 
 	@Override
 	public RestaurantSummaryResponse searchRestaurantSummaries(RestaurantFilterRequest filterRequest) {
-		Slice<Object[]> resultObjects = restaurantRepository.findRestaurantByFilter(
-			filterRequest.getMaxPrice(),
-			filterRequest.getOptions().getFoodTypeName(),
-			filterRequest.computeMySqlPoint(),
-			filterRequest.getMaxDistance(),
-			filterRequest.createPageRequest()
-		);
-		return RestaurantSummaryResponse.of(resultObjects);
+
+		RestaurantSearchCondition searchDto = new RestaurantSearchCondition(filterRequest);
+
+		if (searchDto.getSearchString() == null) {
+			return RestaurantSummaryResponse.ofSummary(
+				restaurantRepository.searchRestaurant(searchDto));
+		}
+		return RestaurantSummaryResponse.ofSummary(
+			restaurantRepository.searchRestaurantContainStringSearch(searchDto));
 	}
 
 	@Override
 	public RestaurantDetailResponse getRestaurantDetail(long restaurantId) {
-		Restaurant restaurant = restaurantRepository.findByIdWithMenus(restaurantId).orElseThrow(NoRecordException::new);
-		return RestaurantDetailResponse.of(restaurant);
+		Restaurant restaurantWithMenus = restaurantRepository.findByIdWithMenus(restaurantId).orElseThrow(NoRecordException::new);
+		Restaurant restaurantWithCategories = restaurantRepository.findByIdWithCategories(restaurantId).orElseThrow(NoRecordException::new);
+		restaurantWithMenus.setCategories(restaurantWithCategories.getCategories());
+		return RestaurantDetailResponse.of(restaurantWithMenus);
 	}
 }
